# flytwo-pro-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FlyTwo Pro Backend API Server",
        "title": "FlyTwo Pro API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "FlyTwo Support",
            "url": "https://flytwo.com/support",
            "email": "support@flytwo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3080",
    "basePath": "/api/v1",
    "paths": {
        "/catmat/import": {
            "post": {
                "description": "Faz upsert dos itens da planilha para a tabela catmat_item",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catmat"
                ],
                "summary": "Importa planilha CATMAT (XLSX)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Arquivo .xlsx com colunas do CATMAT",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ImportResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/catmat/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Pesquisa itens do catálogo CATMAT usando busca textual com filtros opcionais",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catmat"
                ],
                "summary": "Pesquisa itens CATMAT via full-text search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Termo de busca",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Código do grupo (2 dígitos)",
                        "name": "group_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Código da classe (4 dígitos)",
                        "name": "class_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Código do PDM (5 dígitos)",
                        "name": "pdm_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Código NCM",
                        "name": "ncm_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limite de resultados (padrão 50, máximo 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset para paginação (padrão 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados da busca paginados",
                        "schema": {
                            "$ref": "#/definitions/dto.CatmatSearchResponse"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/catser/import": {
            "post": {
                "description": "Faz upsert dos itens da planilha para a tabela catser_item",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catser"
                ],
                "summary": "Importa planilha CATSER (XLSX)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Arquivo .xlsx com colunas do CATSER",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ImportResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/catser/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Pesquisa itens do catálogo CATSER usando busca textual com filtros opcionais",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catser"
                ],
                "summary": "Pesquisa itens CATSER via full-text search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Termo de busca",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Código do grupo",
                        "name": "group_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Código da classe",
                        "name": "class_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Código do serviço",
                        "name": "service_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status (Ativo/Inativo)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limite de resultados (padrão 50, máximo 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset para paginação (padrão 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados da busca paginados",
                        "schema": {
                            "$ref": "#/definitions/dto.CatserSearchResponse"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Authenticate a user and create a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "End the current user session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the profile information of the currently authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "User profile data",
                        "schema": {
                            "$ref": "#/definitions/dto.UserProfileResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/signup": {
            "post": {
                "description": "Register a new user with username, email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User signup information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Validation errors or email/username already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CatmatSearchItem": {
            "type": "object",
            "properties": {
                "class_code": {
                    "type": "integer"
                },
                "class_name": {
                    "type": "string"
                },
                "group_code": {
                    "type": "integer"
                },
                "group_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_code": {
                    "type": "integer"
                },
                "item_description": {
                    "type": "string"
                },
                "ncm_code": {
                    "type": "string"
                },
                "pdm_code": {
                    "type": "integer"
                },
                "pdm_name": {
                    "type": "string"
                },
                "rank": {
                    "type": "number"
                }
            }
        },
        "dto.CatmatSearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CatmatSearchItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.CatserSearchItem": {
            "type": "object",
            "properties": {
                "class_code": {
                    "type": "integer"
                },
                "class_name": {
                    "type": "string"
                },
                "group_code": {
                    "type": "integer"
                },
                "group_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "material_service_type": {
                    "type": "string"
                },
                "rank": {
                    "type": "number"
                },
                "service_code": {
                    "type": "integer"
                },
                "service_description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.CatserSearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CatserSearchItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateUserReq": {
            "type": "object",
            "required": [
                "email",
                "password",
                "user_name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "dto.LoginUserReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.UserProfileResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "services.ImportResult": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.RowError"
                    }
                },
                "rows_read": {
                    "type": "integer"
                },
                "rows_saved": {
                    "type": "integer"
                },
                "rows_skipped": {
                    "type": "integer"
                }
            }
        },
        "services.RowError": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "row": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}