# flytwo-pro-frontend/src/app/api/models/services/import-result.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RowError as ServicesRowError } from '../../models/services/row-error';
export interface ImportResult {
  errors?: Array<ServicesRowError>;
  rows_read?: number;
  rows_saved?: number;
  rows_skipped?: number;
}